requirements.txt
Django==1.11
django-pagedown==1.0.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
